renamed SearchResult to ChatMessage
